event_management_system/requirements.txt
Flask>=2.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
Pillow>=10.1.0
ReportLab>=4.0.0
opencv-python>=4.8.0
face-recognition>=1.3.0
razorpay>=1.3.0
requests>=2.31.0
qrcode>=7.4.0
python-dateutil>=2.8.0
Werkzeug>=2.3.0
